feat(types): add isize primitive + migrate arr.len/index/Range to usize
isize lands as a first-class signed pointer-size primitive (maps to
ptrdiff_t in C, 64-bit i64.const in bytecode/VM). Threaded through every
backend (typecheck, bytecode, VM, C-emit, comptime) and stdlib (Display,
Equals, Hash, Add/Sub/Mul/Div impls in std/core).

Coordinated migration of array length / index / iteration types from
i32/u32 to usize :

- arr.len() now returns usize (typecheck inferField, lower LoweredArrayLen,
  VM array.len pushes i64("usize"), C-emit returns size_t)
- ArrayIter.cursor / .length are usize ; for-in's wrapArrayAsIter emits
  usize literals
- std/iter signatures : count/take/skip/slice/enumerate/zip use usize for
  index/length params and inner counters
- Range[usize] impls Contains+Iterator added in std/core ; inferRange picks
  bound type from first concrete operand (probes upper when lower is
  FreeInt), forInElementType reads the resolved Range[T] struct args
  instead of hardcoding i32 — `for i in 0..<arr.len()` binds i: usize
- Stdlib callers adjusted : std/sort (lo/hi + INSERTION_CUTOFF as
  usize(16)), std/path (normalize cursors), std/json (write_array/object),
  std/collections (BUCKET_COUNT, keys/values cursors), std/numbers
  (to_hex/to_bin rewritten as `for k < n { rev[n-1-k] }` to avoid usize
  underflow on the MSB walk)
- Snippets regex_helpers + selfhost_lexer_basic + example mowitnow updated
  to type their cursors as usize

Bidirectional typing fix in inferGenericFnCall : FreeInt/FreeFloat arg
literals are now repinned in exprTypes after the substitution is built, so
`take(arr, 3)` on `take[T](arr, n: usize)` emits i64.const 3 instead of
i32.const 3 (was a tag-mismatch trap in the VM).

Bytecode patch bumped 0.1.0 → 0.1.1 (additive VAL_TYPES + op-kinds for
isize). VM/native tests : 355 pass. Snapshots regenerated for symbol-id
drift caused by the new core impls. arr.len() == 12 and arr[2] still
compile without explicit casts thanks to FreeInt → usize bidirectional
coercion and the VM's tag-agnostic asIndex helper.

+++ src/lower/passes/block.ts
@@ -311,7 +311,7 @@ function emitArrayDestructure(
     if (leaf.kind === "WildcardBinding") continue;
     const access: LoweredExpr = {
       kind: "LoweredIndex", span: leaf.span, type: elemType,
-      target, index: { kind: "LoweredIntLit", span: leaf.span, type: TY.i32, value: BigInt(i) },
+      target, index: { kind: "LoweredIntLit", span: leaf.span, type: TY.usize, value: BigInt(i) },
     };
     emitLetBindingLeaves(ctx, leaf, access, elemType, leaf.span, out);
   }
@@ -328,10 +328,10 @@ function emitArrayDestructure(
   });
   const idxSym = freshSyntheticSymbol(ctx, "i");
   out.push({
-    kind: "LoweredLet", span, name: idxSym.name, symbol: idxSym, type: TY.i32,
-    value: { kind: "LoweredIntLit", span, type: TY.i32, value: BigInt(restIdx) },
+    kind: "LoweredLet", span, name: idxSym.name, symbol: idxSym, type: TY.usize,
+    value: { kind: "LoweredIntLit", span, type: TY.usize, value: BigInt(restIdx) },
   });
-  const idxRef: LoweredExpr = { kind: "LoweredIdent", span, type: TY.i32, symbol: idxSym };
+  const idxRef: LoweredExpr = { kind: "LoweredIdent", span, type: TY.usize, symbol: idxSym };
   const restRef = readLocal(ctx, restSym, targetType, span);
   const pushStmt: LoweredStmt = {
     kind: "LoweredExprStmt", span,
@@ -343,16 +343,16 @@ function emitArrayDestructure(
   const incStmt: LoweredStmt = {
     kind: "LoweredAssign", span, target: idxRef,
     value: {
-      kind: "LoweredBinary", span, type: TY.i32, op: "add",
+      kind: "LoweredBinary", span, type: TY.usize, op: "add",
       left: idxRef,
-      right: { kind: "LoweredIntLit", span, type: TY.i32, value: 1n },
+      right: { kind: "LoweredIntLit", span, type: TY.usize, value: 1n },
     },
   };
   out.push({
     kind: "LoweredLoop", span, label: null, cond: {
       kind: "LoweredBinary", span, type: TY.bool, op: "lt",
       left: idxRef,
-      right: { kind: "LoweredArrayLen", span, type: TY.i32, target },
+      right: { kind: "LoweredArrayLen", span, type: TY.usize, target },
     },
     body: { kind: "LoweredBlock", span, type: TY.void, stmts: [pushStmt, incStmt], trailing: null },
   });
